@@ -11,9 +11,9 @@ public protocol ObjectScopeProtocol: AnyObject {
 }
 
 /// Basic implementation of ``ObjectScopeProtocol``.
-public class ObjectScope: ObjectScopeProtocol, CustomStringConvertible {
-    public private(set) var description: String
-    private var storageFactory: () -> InstanceStorage
+public final class ObjectScope: ObjectScopeProtocol, CustomStringConvertible, @unchecked Sendable {
+    public let description: String
+    private let storageFactory: () -> InstanceStorage
     private let parent: ObjectScopeProtocol?
 
     /// Instantiates an ``ObjectScope`` with storage factory and description.
@@ -31,7 +31,7 @@ public class ObjectScope: ObjectScopeProtocol, CustomStringConvertible {
         self.parent = parent
     }
 
-    /// Will invoke and return the result of `storageFactory` closure provided during initialisation.
+    /// Will invoke and return the result of `storageFactory` closure provided during initialization.
     public func makeStorage() -> InstanceStorage {
         if let parent = parent {
             return CompositeStorage([storageFactory(), parent.makeStorage()])
@@ -39,4 +39,22 @@ public class ObjectScope: ObjectScopeProtocol, CustomStringConvertible {
             return storageFactory()
         }
     }
+    
+    /// A new instance is always created by the ``Container`` when a type is resolved.
+    /// The instance is not shared.
+    public static let transient = ObjectScope(storageFactory: TransientStorage.init, description: "transient")
+
+    /// Instances are shared only when an object graph is being created,
+    /// otherwise a new instance is created by the ``Container``. This is the default scope.
+    public static let graph = ObjectScope(storageFactory: GraphStorage.init, description: "graph")
+
+    /// An instance provided by the ``Container`` is shared within the ``Container`` and its child `Containers`.
+    public static let container = ObjectScope(storageFactory: PermanentStorage.init, description: "container")
+
+    /// An instance provided by the ``Container`` is shared within the ``Container`` and its child ``Container``s
+    /// as long as there are strong references to given instance. Otherwise new instance is created
+    /// when resolving the type.
+    public static let weak = ObjectScope(storageFactory: WeakStorage.init, description: "weak",
+                                         parent: ObjectScope.graph)
+
 }
